Conflicts:
	data/images/recordsheets/Conventional_Infantry_no_tables.svg
	data/images/recordsheets/Conventional_Infantry_tables.svg
	src/megameklab/com/util/UnitPrintManager.java
	src/megameklab/com/util/UnitUtil.java

+++ src/megameklab/com/ui/BattleArmor/StatusBar.java
@@ -138,7 +138,7 @@ public class StatusBar extends ITab {
         double maxKilos = getBattleArmor().getTrooperWeight();
         double currentKilos;
         int bv = getBattleArmor().calculateBattleValue();
-        int currentCost = (int)Math.round(getBattleArmor().getCost(false));
+        long currentCost = (long) Math.round(getBattleArmor().getCost(false));
 
         testBA = new TestBattleArmor(getBattleArmor(), entityVerifier.baOption,
                 null);
